looks like we are good to merge the status fix. This fix covers: Users added by the legacy system Users added by admin It forces users to complete the registration process before they can search the positions or users or share their profile on other systems.
It also fixes the output issues with the employer profiles.

The email confirm has been added to the settings page, but it still exists individually.

Things to be done, not necessarily associated with this fix, is to remove teh legacy registration page (/register) and remove the confirm email page, since the confirm email button is now on the settings page

commented out debug, it seems to be working fine now